requirements.txt
apache-beam==2.12.0
avro==1.9.0
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
crcmod==1.7
dill==0.2.9
docopt==0.6.2
enum34==1.1.6
fastavro==0.21.24
fasteners==0.15
funcsigs==1.0.2
future==0.17.1
futures==3.2.0
google-api-core==1.11.1
google-apitools==0.5.26
google-auth==1.6.3
google-cloud-bigquery==1.6.1
google-cloud-bigtable==0.31.1
google-cloud-core==1.0.1
google-cloud-pubsub==0.39.0
google-cloud-storage==1.16.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
googledatastore==7.0.2
grpc-google-iam-v1==0.11.4
grpcio==1.21.1
hdfs==2.5.2
httplib2==0.11.3
idna==2.8
mock==2.0.0
monotonic==1.5
numpy==1.16.4
oauth2client==3.0.0
pbr==5.2.1
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.8.0
pyarrow==0.11.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pydot==1.2.4
pyparsing==2.4.0
pytz==2019.1
PyVCF==0.6.8
PyYAML==3.13
requests==2.22.0
rsa==4.0
six==1.12.0
typing==3.6.6
urllib3==1.25.3
